# AST-60: [experimental] generate typescript interfaces from JSON schema