--- a/frontend/webpack/json-schema/compile.js
+++ b/frontend/webpack/json-schema/compile.js
@@ -1,0 +1,9 @@
+/* eslint-env es6 */
+const fs = require('fs');
+const { compileFromFile } = require('json-schema-to-typescript');
+const colors = require('colors');
+
+module.exports = async function compile(schemaPath, typescriptPath) {
+  const content = await compileFromFile(schemaPath);
+  fs.writeFileSync(typescriptPath, content);
+};