Original PR : https://github.com/akeneo/pim-community-dev/pull/11030

**Description (for Contributor and Core Developer)**

This is experimental.

We want to define JSON schema files for validating API responses.
The validation will be then simplified, when used in the fetchers, with a library like https://github.com/epoberezkin/ajv
We also want to generate non-versioned typescript interfaces from the same JSON schema.

This PR includes 2 webpack changes:
- a **resolver plugin**: when one of the generated typescript interface is imported somewhere in the code, the resolver will ensure this generated file exists instead of throwing a "module not found" error.
- a **loader**: every time a `.schema.json` file is imported somewhere in the code, the loader will generate its typescript interface counterpart. This is really useful for webpack watch.

Since we are hooking into the resolver, I've been doing some performance checks to ensure we are not slowing down webpack too much.

```
existing webpack config + no cache
time for i in {1..10}; do echo ${i} && rm -rf web/cache && run node yarn run webpack-dev > /dev/null; done
7m16,659s

resolver plugin enabled + no cache
time for i in {1..10}; do echo ${i} && rm -rf web/cache && run node yarn run webpack-dev > /dev/null; done
8m7,952s

existing webpack config + cache
time for i in {1..10}; do echo ${i} && run node yarn run webpack-dev > /dev/null; done
6m56,124s

resolver plugin enabled + cache
time for i in {1..10}; do echo ${i} && run node yarn run webpack-dev > /dev/null; done
6m56,998s

resolver plugin enabled + cache + force recompile
time for i in {1..10}; do echo ${i} && rm -f web/js/json-schema/model/family.ts && run node yarn run webpack-dev > /dev/null; done
6m53,628s
```

There is an average increase of 11% when removing the cache. Otherwise, the difference is unnoticeable.

**Definition Of Done (for Core Developer only)**

| Q                                 | A
| --------------------------------- | ---
| Added Specs                       | Todo
| Added legacy Behats               | Todo
| Added acceptance tests            | Todo
| Added integration tests           | Todo
| Changelog updated                 | Todo
| Review and 2 GTM                  | Todo
| Micro Demo to the PO (Story only) | Todo
| Migration script                  | -
| Tech Doc                          | -